refactor: 파일명 및 변수 형식 변경
