+++ src/styles/styled.d.ts
@@ -1,11 +1,11 @@
 import 'styled-components'
-import { color, font, media } from './theme'
+import { COLOR, FONT, MEDIA } from './theme'
 import {} from 'styled-components/cssprop'
 
 declare module 'styled-components' {
   export interface DefaultTheme {
-    color: typeof color
-    font: typeof font
-    media: typeof media
+    color: typeof COLOR
+    font: typeof FONT
+    media: typeof MEDIA
   }
 }
